delete dead output-writer + frame-renderer; isolate RefCounter
OutputWriter / StdoutWriter / BufferWriter / FrameRenderer were unused
across the codebase. Their goal (test-buffer + diff-render) is subsumed
by RenderSurface and DefaultCompositor's compositor:write event, plus
the surface-routing migrations from earlier in this branch.

The only live consumer was shell.ts using RefCounter, which is a
generic utility unrelated to output writing. Moved to ref-counter.ts.

Net: -152 lines, +9 lines.
